撤回消息修复

+++ src/main/java/Color_yr/ColorMirai/Pack/FromPlugin/ReCallMessagePack.java
@@ -5,5 +5,5 @@ package Color_yr.ColorMirai.Pack.FromPlugin;
 id：消息Id
  */
 public class ReCallMessagePack {
-    public long id;
+    public int id;
 }
